+++ library/src/extra/csrgemm_device.h
@@ -28,7 +28,7 @@
 
 namespace rocsparse
 {
-    template <unsigned int BLOCKSIZE, typename I, typename J>
+    template <uint32_t BLOCKSIZE, typename I, typename J>
     ROCSPARSE_KERNEL(BLOCKSIZE)
     void csrgemm_set_base(I size, J* __restrict__ out, rocsparse_index_base idx_base_out)
     {
@@ -42,7 +42,7 @@ namespace rocsparse
     }
 
     // Decrement
-    template <unsigned int BLOCKSIZE, typename I>
+    template <uint32_t BLOCKSIZE, typename I>
     ROCSPARSE_KERNEL(BLOCKSIZE)
     void csrgemm_index_base(I* nnz)
     {
@@ -50,7 +50,7 @@ namespace rocsparse
     }
 
     // Copy an array
-    template <unsigned int BLOCKSIZE, typename I, typename J>
+    template <uint32_t BLOCKSIZE, typename I, typename J>
     ROCSPARSE_KERNEL(BLOCKSIZE)
     void csrgemm_copy(I size,
                       const J* __restrict__ in,
@@ -69,7 +69,7 @@ namespace rocsparse
     }
 
     // Copy and scale an array
-    template <unsigned int BLOCKSIZE, typename I, typename T>
+    template <uint32_t BLOCKSIZE, typename I, typename T>
     ROCSPARSE_DEVICE_ILF void csrgemm_copy_scale_device(I size, T alpha, const T* in, T* out)
     {
         I idx = hipBlockIdx_x * BLOCKSIZE + hipThreadIdx_x;
@@ -83,7 +83,7 @@ namespace rocsparse
     }
 
     // Compute number of intermediate products of each row
-    template <unsigned int BLOCKSIZE, unsigned int WFSIZE, typename I, typename J>
+    template <uint32_t BLOCKSIZE, uint32_t WFSIZE, typename I, typename J>
     ROCSPARSE_KERNEL(BLOCKSIZE)
     void csrgemm_intermediate_products(J m,
                                        const I* __restrict__ csr_row_ptr_A,
@@ -145,24 +145,24 @@ namespace rocsparse
         }
     }
 
-    template <unsigned int BLOCKSIZE, unsigned int GROUPS, typename I>
+    template <uint32_t BLOCKSIZE, uint32_t GROUPS, typename I>
     ROCSPARSE_DEVICE_ILF void csrgemm_group_reduce(int tid, I* __restrict__ data)
     {
         // clang-format off
-    if(BLOCKSIZE > 512 && tid < 512) for(unsigned int i = 0; i < GROUPS; ++i) data[tid * GROUPS + i] += data[(tid + 512) * GROUPS + i]; __syncthreads();
-    if(BLOCKSIZE > 256 && tid < 256) for(unsigned int i = 0; i < GROUPS; ++i) data[tid * GROUPS + i] += data[(tid + 256) * GROUPS + i]; __syncthreads();
-    if(BLOCKSIZE > 128 && tid < 128) for(unsigned int i = 0; i < GROUPS; ++i) data[tid * GROUPS + i] += data[(tid + 128) * GROUPS + i]; __syncthreads();
-    if(BLOCKSIZE >  64 && tid <  64) for(unsigned int i = 0; i < GROUPS; ++i) data[tid * GROUPS + i] += data[(tid +  64) * GROUPS + i]; __syncthreads();
-    if(BLOCKSIZE >  32 && tid <  32) for(unsigned int i = 0; i < GROUPS; ++i) data[tid * GROUPS + i] += data[(tid +  32) * GROUPS + i]; __syncthreads();
-    if(BLOCKSIZE >  16 && tid <  16) for(unsigned int i = 0; i < GROUPS; ++i) data[tid * GROUPS + i] += data[(tid +  16) * GROUPS + i]; __syncthreads();
-    if(BLOCKSIZE >   8 && tid <   8) for(unsigned int i = 0; i < GROUPS; ++i) data[tid * GROUPS + i] += data[(tid +   8) * GROUPS + i]; __syncthreads();
-    if(BLOCKSIZE >   4 && tid <   4) for(unsigned int i = 0; i < GROUPS; ++i) data[tid * GROUPS + i] += data[(tid +   4) * GROUPS + i]; __syncthreads();
-    if(BLOCKSIZE >   2 && tid <   2) for(unsigned int i = 0; i < GROUPS; ++i) data[tid * GROUPS + i] += data[(tid +   2) * GROUPS + i]; __syncthreads();
-    if(BLOCKSIZE >   1 && tid <   1) for(unsigned int i = 0; i < GROUPS; ++i) data[tid * GROUPS + i] += data[(tid +   1) * GROUPS + i]; __syncthreads();
+    if(BLOCKSIZE > 512 && tid < 512) for(uint32_t i = 0; i < GROUPS; ++i) data[tid * GROUPS + i] += data[(tid + 512) * GROUPS + i]; __syncthreads();
+    if(BLOCKSIZE > 256 && tid < 256) for(uint32_t i = 0; i < GROUPS; ++i) data[tid * GROUPS + i] += data[(tid + 256) * GROUPS + i]; __syncthreads();
+    if(BLOCKSIZE > 128 && tid < 128) for(uint32_t i = 0; i < GROUPS; ++i) data[tid * GROUPS + i] += data[(tid + 128) * GROUPS + i]; __syncthreads();
+    if(BLOCKSIZE >  64 && tid <  64) for(uint32_t i = 0; i < GROUPS; ++i) data[tid * GROUPS + i] += data[(tid +  64) * GROUPS + i]; __syncthreads();
+    if(BLOCKSIZE >  32 && tid <  32) for(uint32_t i = 0; i < GROUPS; ++i) data[tid * GROUPS + i] += data[(tid +  32) * GROUPS + i]; __syncthreads();
+    if(BLOCKSIZE >  16 && tid <  16) for(uint32_t i = 0; i < GROUPS; ++i) data[tid * GROUPS + i] += data[(tid +  16) * GROUPS + i]; __syncthreads();
+    if(BLOCKSIZE >   8 && tid <   8) for(uint32_t i = 0; i < GROUPS; ++i) data[tid * GROUPS + i] += data[(tid +   8) * GROUPS + i]; __syncthreads();
+    if(BLOCKSIZE >   4 && tid <   4) for(uint32_t i = 0; i < GROUPS; ++i) data[tid * GROUPS + i] += data[(tid +   4) * GROUPS + i]; __syncthreads();
+    if(BLOCKSIZE >   2 && tid <   2) for(uint32_t i = 0; i < GROUPS; ++i) data[tid * GROUPS + i] += data[(tid +   2) * GROUPS + i]; __syncthreads();
+    if(BLOCKSIZE >   1 && tid <   1) for(uint32_t i = 0; i < GROUPS; ++i) data[tid * GROUPS + i] += data[(tid +   1) * GROUPS + i]; __syncthreads();
         // clang-format on
     }
 
-    template <unsigned int BLOCKSIZE, unsigned int GROUPS, typename I, typename J>
+    template <uint32_t BLOCKSIZE, uint32_t GROUPS, typename I, typename J>
     ROCSPARSE_KERNEL(BLOCKSIZE)
     void csrgemm_group_reduce_part1(J m, I* __restrict__ int_prod, J* __restrict__ group_size)
     {
@@ -172,7 +172,7 @@ namespace rocsparse
         __shared__ J sdata[BLOCKSIZE * GROUPS];
 
         // Initialize shared memory
-        for(unsigned int i = 0; i < GROUPS; ++i)
+        for(uint32_t i = 0; i < GROUPS; ++i)
         {
             sdata[hipThreadIdx_x * GROUPS + i] = 0;
         }
@@ -209,7 +209,7 @@ namespace rocsparse
         }
     }
 
-    template <unsigned int BLOCKSIZE, unsigned int GROUPS, bool CPLX, typename I, typename J>
+    template <uint32_t BLOCKSIZE, uint32_t GROUPS, bool CPLX, typename I, typename J>
     ROCSPARSE_KERNEL(BLOCKSIZE)
     void csrgemm_group_reduce_part2(J m,
                                     const I* __restrict__ csr_row_ptr,
@@ -222,7 +222,7 @@ namespace rocsparse
         __shared__ J sdata[BLOCKSIZE * GROUPS];
 
         // Initialize shared memory
-        for(unsigned int i = 0; i < GROUPS; ++i)
+        for(uint32_t i = 0; i < GROUPS; ++i)
         {
             sdata[hipThreadIdx_x * GROUPS + i] = 0;
         }
@@ -261,7 +261,7 @@ namespace rocsparse
         }
     }
 
-    template <unsigned int BLOCKSIZE, unsigned int GROUPS, typename I>
+    template <uint32_t BLOCKSIZE, uint32_t GROUPS, typename I>
     ROCSPARSE_KERNEL(BLOCKSIZE)
     void csrgemm_group_reduce_part3(I* __restrict__ group_size)
     {
@@ -269,7 +269,7 @@ namespace rocsparse
         __shared__ I sdata[BLOCKSIZE * GROUPS];
 
         // Copy global data to shared memory
-        for(unsigned int i = hipThreadIdx_x; i < BLOCKSIZE * GROUPS; i += BLOCKSIZE)
+        for(uint32_t i = hipThreadIdx_x; i < BLOCKSIZE * GROUPS; i += BLOCKSIZE)
         {
             sdata[i] = group_size[i];
         }
@@ -287,7 +287,7 @@ namespace rocsparse
         }
     }
 
-    template <unsigned int BLOCKSIZE, typename I, typename J>
+    template <uint32_t BLOCKSIZE, typename I, typename J>
     ROCSPARSE_KERNEL(BLOCKSIZE)
     void csrgemm_max_row_nnz_part1(J m,
                                    const I* __restrict__ csr_row_ptr,
@@ -324,7 +324,7 @@ namespace rocsparse
         }
     }
 
-    template <unsigned int BLOCKSIZE, typename I>
+    template <uint32_t BLOCKSIZE, typename I>
     ROCSPARSE_KERNEL(BLOCKSIZE)
     void csrgemm_max_row_nnz_part2(I* __restrict__ workspace)
     {
@@ -349,7 +349,7 @@ namespace rocsparse
 
     // Hash operation to insert key into hash table
     // Returns true if key has been added
-    template <unsigned int HASHVAL, unsigned int HASHSIZE, typename I>
+    template <uint32_t HASHVAL, uint32_t HASHSIZE, typename I>
     ROCSPARSE_DEVICE_ILF bool insert_key(I key, I* __restrict__ table)
     {
         // Compute hash
@@ -383,7 +383,7 @@ namespace rocsparse
     }
 
     // Hash operation to insert pair into hash table
-    template <unsigned int HASHVAL, unsigned int HASHSIZE, typename I, typename T>
+    template <uint32_t HASHVAL, uint32_t HASHSIZE, typename I, typename T>
     ROCSPARSE_DEVICE_ILF void
         insert_pair(I key, T val, I* __restrict__ table, T* __restrict__ data, I empty)
     {
@@ -418,10 +418,10 @@ namespace rocsparse
     }
 
     // Compute non-zero entries per row, where each row is processed by a single wavefront
-    template <unsigned int BLOCKSIZE,
-              unsigned int WFSIZE,
-              unsigned int HASHSIZE,
-              unsigned int HASHVAL,
+    template <uint32_t BLOCKSIZE,
+              uint32_t WFSIZE,
+              uint32_t HASHSIZE,
+              uint32_t HASHVAL,
               typename I,
               typename J>
     ROCSPARSE_KERNEL(BLOCKSIZE)
@@ -456,7 +456,7 @@ namespace rocsparse
         J* table = &stable[wid * HASHSIZE];
 
         // Initialize hash table
-        for(unsigned int i = lid; i < HASHSIZE; i += WFSIZE)
+        for(uint32_t i = lid; i < HASHSIZE; i += WFSIZE)
         {
             table[i] = -1;
         }
@@ -528,10 +528,10 @@ namespace rocsparse
     }
 
     // Compute non-zero entries per row, where each row is processed by a single block
-    template <unsigned int BLOCKSIZE,
-              unsigned int WFSIZE,
-              unsigned int HASHSIZE,
-              unsigned int HASHVAL,
+    template <uint32_t BLOCKSIZE,
+              uint32_t WFSIZE,
+              uint32_t HASHSIZE,
+              uint32_t HASHVAL,
               typename I,
               typename J>
     ROCSPARSE_KERNEL(BLOCKSIZE)
@@ -562,7 +562,7 @@ namespace rocsparse
         __shared__ J table[HASHSIZE];
 
         // Initialize hash table
-        for(unsigned int i = hipThreadIdx_x; i < HASHSIZE; i += BLOCKSIZE)
+        for(uint32_t i = hipThreadIdx_x; i < HASHSIZE; i += BLOCKSIZE)
         {
             table[i] = -1;
         }
@@ -646,11 +646,7 @@ namespace rocsparse
     // Splitting row into several chunks such that we can use shared memory to store whether
     // a column index is populated or not.
     // Each row has at least 8193 intermediate products to compute.
-    template <unsigned int BLOCKSIZE,
-              unsigned int WFSIZE,
-              unsigned int CHUNKSIZE,
-              typename I,
-              typename J>
+    template <uint32_t BLOCKSIZE, uint32_t WFSIZE, uint32_t CHUNKSIZE, typename I, typename J>
     ROCSPARSE_KERNEL(BLOCKSIZE)
     void csrgemm_nnz_block_per_row_multipass(J n,
                                              const J* __restrict__ offset,
@@ -706,7 +702,7 @@ namespace rocsparse
         while(chunk_begin < n)
         {
             // Initialize row nnz table
-            for(unsigned int i = hipThreadIdx_x; i < CHUNKSIZE; i += BLOCKSIZE)
+            for(uint32_t i = hipThreadIdx_x; i < CHUNKSIZE; i += BLOCKSIZE)
             {
                 table[i] = false;
             }
@@ -820,7 +816,7 @@ namespace rocsparse
 
             // Each thread loads its entry for the current chunk
             J chunk_nnz = 0;
-            for(unsigned int i = hipThreadIdx_x; i < CHUNKSIZE; i += BLOCKSIZE)
+            for(uint32_t i = hipThreadIdx_x; i < CHUNKSIZE; i += BLOCKSIZE)
             {
                 chunk_nnz += (table[i] == true) ? 1 : 0;
             }
@@ -854,10 +850,10 @@ namespace rocsparse
     }
 
     // Compute column entries and accumulate values, where each row is processed by a single wavefront
-    template <unsigned int BLOCKSIZE,
-              unsigned int WFSIZE,
-              unsigned int HASHSIZE,
-              unsigned int HASHVAL,
+    template <uint32_t BLOCKSIZE,
+              uint32_t WFSIZE,
+              uint32_t HASHSIZE,
+              uint32_t HASHVAL,
               typename I,
               typename J,
               typename T>
@@ -903,7 +899,7 @@ namespace rocsparse
         T* data  = &sdata[wid * HASHSIZE];
 
         // Initialize hash table
-        for(unsigned int i = lid; i < HASHSIZE; i += WFSIZE)
+        for(uint32_t i = lid; i < HASHSIZE; i += WFSIZE)
         {
             table[i] = nk;
             data[i]  = static_cast<T>(0);
@@ -971,7 +967,7 @@ namespace rocsparse
         I row_begin_C = csr_row_ptr_C[row] - idx_base_C;
 
         // Loop over hash table
-        for(unsigned int i = lid; i < HASHSIZE; i += WFSIZE)
+        for(uint32_t i = lid; i < HASHSIZE; i += WFSIZE)
         {
             // Get column from hash table to fill it into C
             J col_C = table[i];
@@ -986,7 +982,7 @@ namespace rocsparse
             I idx_C = row_begin_C;
 
             // Initialize index into hash table
-            unsigned int hash_idx = 0;
+            uint32_t hash_idx = 0;
 
             // Loop through hash table to find the (sorted) index into C for the
             // current column index
@@ -1011,10 +1007,10 @@ namespace rocsparse
     }
 
     // Compute column entries and accumulate values, where each row is processed by a single block
-    template <unsigned int BLOCKSIZE,
-              unsigned int WFSIZE,
-              unsigned int HASHSIZE,
-              unsigned int HASHVAL,
+    template <uint32_t BLOCKSIZE,
+              uint32_t WFSIZE,
+              uint32_t HASHSIZE,
+              uint32_t HASHVAL,
               typename I,
               typename J,
               typename T>
@@ -1052,7 +1048,7 @@ namespace rocsparse
         __shared__ T data[HASHSIZE];
 
         // Initialize hash table
-        for(unsigned int i = hipThreadIdx_x; i < HASHSIZE; i += BLOCKSIZE)
+        for(uint32_t i = hipThreadIdx_x; i < HASHSIZE; i += BLOCKSIZE)
         {
             table[i] = nk;
             data[i]  = static_cast<T>(0);
@@ -1118,7 +1114,7 @@ namespace rocsparse
         J hash_offset = 0;
 
         // Loop over the hash table and do the compression
-        for(unsigned int i = hipThreadIdx_x; i < HASHSIZE; i += BLOCKSIZE)
+        for(uint32_t i = hipThreadIdx_x; i < HASHSIZE; i += BLOCKSIZE)
         {
             // Get column and value from hash table
             J col_C = table[i];
@@ -1129,7 +1125,7 @@ namespace rocsparse
 
             // Each thread obtains a bit mask of all wavefront-wide non-zero entries
             // to compute its wavefront-wide non-zero offset
-            unsigned long long mask = __ballot(has_nnz);
+            uint64_t mask = __ballot(has_nnz);
 
             // The number of bits set to 1 is the amount of wavefront-wide non-zeros
             int nnz = __popcll(mask);
@@ -1137,8 +1133,7 @@ namespace rocsparse
             // Obtain the lane mask, where all bits lesser equal the lane id are set to 1
             // e.g. for lane id 7, lanemask_le = 0b11111111
             // HIP implements only __lanemask_lt() unfortunately ...
-            unsigned long long lanemask_le
-                = UINT64_MAX >> (sizeof(unsigned long long) * CHAR_BIT - (__lane_id() + 1));
+            uint64_t lanemask_le = UINT64_MAX >> (sizeof(uint64_t) * CHAR_BIT - (__lane_id() + 1));
 
             // Compute the intra wavefront offset of the lane id by bitwise AND with the lane mask
             int offset = __popcll(lanemask_le & mask);
@@ -1154,7 +1149,7 @@ namespace rocsparse
             __syncthreads();
 
             // Each thread accumulates the offset of all previous wavefronts to obtain its offset
-            for(unsigned int j = 1; j < BLOCKSIZE / warpSize; ++j)
+            for(uint32_t j = 1; j < BLOCKSIZE / warpSize; ++j)
             {
                 if(hipThreadIdx_x >= j * warpSize)
                 {
@@ -1222,9 +1217,9 @@ namespace rocsparse
     // block. Splitting row into several chunks such that we can use shared memory to store
     // whether a column index is populated or not. Each row has at least 4097 non-zero
     // entries to compute.
-    template <unsigned int BLOCKSIZE,
-              unsigned int WFSIZE,
-              unsigned int CHUNKSIZE,
+    template <uint32_t BLOCKSIZE,
+              uint32_t WFSIZE,
+              uint32_t CHUNKSIZE,
               typename I,
               typename J,
               typename T>
@@ -1286,7 +1281,7 @@ namespace rocsparse
         while(chunk_begin < n)
         {
             // Initialize row nnz table and accumulator
-            for(unsigned int i = hipThreadIdx_x; i < CHUNKSIZE; i += BLOCKSIZE)
+            for(uint32_t i = hipThreadIdx_x; i < CHUNKSIZE; i += BLOCKSIZE)
             {
                 table[i] = 0;
                 data[i]  = static_cast<T>(0);
@@ -1416,7 +1411,7 @@ namespace rocsparse
             // "Pseudo compress" the table array such that we can copy the values over into C
             // In fact, we do an exclusive scan to obtain the index where each non-zero has
             // to be copied to
-            for(unsigned int i = hipThreadIdx_x; i < CHUNKSIZE; i += BLOCKSIZE)
+            for(uint32_t i = hipThreadIdx_x; i < CHUNKSIZE; i += BLOCKSIZE)
             {
                 // Each thread loads its marker and value to know whether it has to process a
                 // non-zero entry or not
@@ -1425,7 +1420,7 @@ namespace rocsparse
 
                 // Each thread obtains a bit mask of all wavefront-wide non-zero entries
                 // to compute its wavefront-wide non-zero offset in C
-                unsigned long long mask = __ballot(has_nnz == true);
+                uint64_t mask = __ballot(has_nnz == true);
 
                 // The number of bits set to 1 is the amount of wavefront-wide non-zeros
                 int nnz = __popcll(mask);
@@ -1433,8 +1428,8 @@ namespace rocsparse
                 // Obtain the lane mask, where all bits lesser equal the lane id are set to 1
                 // e.g. for lane id 7, lanemask_le = 0b11111111
                 // HIP implements only __lanemask_lt() unfortunately ...
-                unsigned long long lanemask_le
-                    = UINT64_MAX >> (sizeof(unsigned long long) * CHAR_BIT - (__lane_id() + 1));
+                uint64_t lanemask_le
+                    = UINT64_MAX >> (sizeof(uint64_t) * CHAR_BIT - (__lane_id() + 1));
 
                 // Compute the intra wavefront offset of the lane id by bitwise AND with the lane mask
                 int offset = __popcll(lanemask_le & mask);
@@ -1451,7 +1446,7 @@ namespace rocsparse
 
                 // Each thread accumulates the offset of all previous wavefronts to obtain its
                 // offset into C
-                for(unsigned int j = 1; j < BLOCKSIZE / warpSize; ++j)
+                for(uint32_t j = 1; j < BLOCKSIZE / warpSize; ++j)
                 {
                     if(hipThreadIdx_x >= j * warpSize)
                     {
